split Relay into traits
